fixed const_iterator

+++ containers/containers_test/srcs/map/rite_arrow.cpp
@@ -1,4 +1,4 @@
-#include "common.hpp"
+ #include "common.hpp"
 #include <list>
 
 #define T1 float
@@ -10,7 +10,7 @@ int		main(void)
 	std::list<T3> lst;
 	unsigned int lst_size = 5;
 	for (unsigned int i = 0; i < lst_size; ++i)
-		lst.push_back(T3(2.5 - i, (i + 1) * 7));
+		lst.push_back(T3(static_cast<float>(2.5 - i), (i + 1) * 7));
 
 	TESTED_NAMESPACE::map<T1, T2> mp(lst.begin(), lst.end());
 	TESTED_NAMESPACE::map<T1, T2>::reverse_iterator it(mp.rbegin());
